Windows doesn't like symlinks
